Rename plugin folders to valid Python identifiers. Output archives maintain the original naming
reason: Python identifiers must not start with a number and may not
contain dashes and periods. Using valid identifier names allow for them to
be treated as modules and allows for the use of pytest, mypy and other tools
that do discovery by module. It's the first step to proper testing